Move coin selector to own file and add to DI
