+++ src/Stratis.Features.FederatedPeg/FedMultiSigManualWithdrawalTransactionBuilder.cs
@@ -5,45 +5,11 @@ using NBitcoin;
 using Stratis.Bitcoin.Features.Wallet;
 using Stratis.Bitcoin.Features.Wallet.Interfaces;
 using Stratis.Features.FederatedPeg.Interfaces;
-using Stratis.Features.FederatedPeg.Wallet;
 using Recipient = Stratis.Features.FederatedPeg.Wallet.Recipient;
-using TransactionBuildContext = Stratis.Features.FederatedPeg.Wallet.TransactionBuildContext;
 using UnspentOutputReference = Stratis.Features.FederatedPeg.Wallet.UnspentOutputReference;
 
 namespace Stratis.Features.FederatedPeg
 {
-    public interface IMultisigCoinSelector
-    {
-        (List<Coin>, List<UnspentOutputReference>) SelectCoins(List<Recipient> recipients);
-    }
-
-    /// <summary>
-    /// Used to wrap a static method with a class that can be mocked for testing.
-    /// </summary>
-    public class MultisigCoinSelector : IMultisigCoinSelector
-    {
-        private readonly Network network;
-
-        private readonly IFederatedPegSettings settings;
-        private readonly IFederationWalletManager walletManager;
-
-        public MultisigCoinSelector(Network network, IFederatedPegSettings settings, IFederationWalletManager walletManager)
-        {
-            this.network = network;
-            this.settings = settings;
-            this.walletManager = walletManager;
-        }
-
-        public (List<Coin>, List<UnspentOutputReference>) SelectCoins(List<Recipient> recipients)
-        {
-            // FederationWalletTransactionHandler only supports signing with a single key - the fed wallet key - so we don't use it to build the transaction.
-            // However we still want to use it to determine what coins we need, so hack this together here to pass in to FederationWalletTransactionHandler.DetermineCoins.
-            var multiSigContext = new TransactionBuildContext(recipients);
-
-            return FederationWalletTransactionHandler.DetermineCoins(this.walletManager, this.network, multiSigContext, this.settings);
-        }
-    }
-
     /// <summary>
     /// Uses this node's federation wallet and the current network fee policy to build withdrawal transactions for the federation multisig.
     /// </summary>
